Work in progress: Host ids instead of peer addresses.
This work represents an attempt to separate the tight binding of IP address
as the primary identifier inside this library.
It's DOOMed. Because cassandra's Peers table doesn't have port,
and peers are free to join the ring later, we must be able to get to the ip address
that cassandra *thinks* it's on, for better or worse, and all cluster nodes must be on the same port.